# apps/api/src/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Telecom FieldOps API
  version: 1.0.0
  description: Security-focused API for authentication, user management, roles and audit-ready operations.
servers:
  - url: http://localhost:3000/api/v1
...
components:
  securitySchemes:
    OAuth2Password:
      type: oauth2
      flows:
        password:
          tokenUrl: /api/v1/auth/login
          scopes: {}
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, detail, instance, correlationId]
      properties:
        type:
          type: string
          example: urn:telecom:error:validation
        title:
          type: string
          example: Validation Error
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Request body validation failed.
        instance:
          type: string
          example: /api/v1/users
        correlationId:
          type: string
          example: c123
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    Product:
      type: object
      required: [id, name, category, isSerialized]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string, enum: [ROUTER, MODEM, ONT, STB, ANTENNA, CABLE, PHONE, TABLET, LAPTOP, SIM] }
        isSerialized: { type: boolean }
    Branch:
      type: object
      required: [id, name, isMain]
      properties:
        id: { type: string }
        name: { type: string }
        isMain: { type: boolean }
    InventoryItem:
      type: object
      required: [id, branchId, productId, qtyAvailable, qtyReserved]
      properties:
        id: { type: string }
        branchId: { type: string }
        productId: { type: string }
        qtyAvailable: { type: integer }
        qtyReserved: { type: integer }
        updatedAt: { type: string, format: date-time, nullable: true }
    ReserveInventoryRequest:
      type: object
      required: [workOrderId, branchId, items]
      properties:
        workOrderId: { type: string }
        branchId: { type: string }
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, qty]
            properties:
              productId: { type: string }
              qty: { type: integer, minimum: 1 }
    ReservationRecord:
      type: object
      required: [workOrderId, branchId, items, reservedAt]
      properties:
        workOrderId: { type: string }
        branchId: { type: string }
        items:
          type: array
          items:
            type: object
            required: [productId, qty]
            properties:
              productId: { type: string }
              qty: { type: integer }
        reservedAt: { type: string, format: date-time }
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
security:
  - BearerAuth: []
paths:
...
  /inventory:
    get:
      tags: [Inventory]
      parameters:
        - name: branchId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: inventory list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InventoryItem" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /inventory/branches:
    get:
      tags: [Inventory]
      responses:
        "200":
          description: branches list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Branch" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /inventory/products:
    get:
      tags: [Inventory]
      responses:
        "200":
          description: products available for inventory operations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /inventory/reservations:
    post:
      tags: [Inventory]
      summary: Reserve inventory for a work order request
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReserveInventoryRequest" }
      responses:
        "201":
          description: reservation created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReservationRecord" }
        "400":
          description: validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
        "409":
          description: insufficient stock / conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /inventory/reservations/{workOrderId}:
    delete:
      tags: [Inventory]
      summary: Release reservation for a work order
      parameters:
        - name: workOrderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: reservation released
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReservationRecord" }
        "404":
          description: reservation not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
        "400":
          description: invalid request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...